Add files via upload
DirGraph is a data structure to store the tree graph and corresponding equations. 
Power_eq's_ar populates the equations in a dirgraph object given a graph object and matrix of impedances. (can be modified to only accept the impedance matrix).
elimVars generates a univariate polynomials for each child of the root that can be used to find all the solutions.